move code to src and __tests__ dir
