+++ __tests__/main.test.ts
@@ -1,4 +1,4 @@
-import {sum} from './main';
+import {sum} from '../src/main';
 
 test('adds 1 + 2 to equal 3', () => {
   expect(sum(1, 2)).toBe(3);
